[refactor] Move evaluator package to pinot-common
