+++ pinot-common/src/test/java/org/apache/pinot/common/evaluator/InbuiltFunctionEvaluatorTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.segment.local.evaluator;
+package org.apache.pinot.common.evaluator;
 
 import java.util.Collections;
 import org.apache.pinot.common.function.FunctionUtils;
