+++ pinot-core/src/test/java/org/apache/pinot/core/data/function/LogicalFunctionsTest.java
@@ -21,7 +21,7 @@ package org.apache.pinot.core.data.function;
 import com.google.common.collect.Lists;
 import java.util.ArrayList;
 import java.util.List;
-import org.apache.pinot.segment.local.evaluator.InbuiltFunctionEvaluator;
+import org.apache.pinot.common.evaluator.InbuiltFunctionEvaluator;
 import org.apache.pinot.spi.data.readers.GenericRow;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
