+++ pinot-udf-test/src/main/java/org/apache/pinot/udf/test/scenarios/ExpressionTransformerTestScenario.java
@@ -21,10 +21,10 @@ package org.apache.pinot.udf.test.scenarios;
 import com.google.auto.service.AutoService;
 import java.util.HashMap;
 import java.util.Map;
+import org.apache.pinot.common.evaluator.FunctionEvaluatorFactory;
 import org.apache.pinot.core.udf.Udf;
 import org.apache.pinot.core.udf.UdfExample;
 import org.apache.pinot.core.udf.UdfSignature;
-import org.apache.pinot.segment.local.evaluator.FunctionEvaluatorFactory;
 import org.apache.pinot.segment.spi.function.FunctionEvaluator;
 import org.apache.pinot.spi.data.readers.GenericRow;
 import org.apache.pinot.udf.test.PinotFunctionEnvGenerator;
